Omitting inactive nodes [skip ci]

+++ actions/cleanupSystem.js
@@ -16,7 +16,13 @@ module.exports = async function executeCleanup(proxy, config) {
     config['cleanups'] = config['cleanups'] ? config['cleanups'] : {};
     const entries = [];
     for (const node of nodes) {
-        await proxy.executeCommand_N(node.hostname, "docker system prune -f")
+        if(node.availability == "Down")
+        {
+            log.info(`escaping node for being inactive: '${node.hostname}'`);
+        }
+        else
+        {
+            await proxy.executeCommand_N(node.hostname, "docker system prune -f")
             .then(response => {
                 const filteredInput = response.split(/\r\n|\r|\n/g).filter(it => it.startsWith("Total reclaimed space:"))[0];
                 if (filteredInput) {
@@ -30,6 +36,7 @@ module.exports = async function executeCleanup(proxy, config) {
                 entries.push(entry);
             })
             .catch(error => log.warn(`could not prune node '${node.hostname}'`, error))
+        }
     }
     config['cleanups'][env] = entries;
